OAuth2 support
The WG11 Security Requirements v04.00.01 defined the API should
following the OAuth2.0, and the O2 IMS API not compliance with
the OAuth2.
This commit will support OAuth2.0 with O2 IMS API. The OAuth2
server based on a 3rd-party, and register the O2 IMS to OAuth
server to enable the OAuth funcational.
The WG11 Security Requirements v04.00.01 specify that the API should
follow OAuth 2.0. However, the O2 IMS API is not compliant with OAuth
2.0.
This commit will add OAuth 2.0 support to the O2 IMS API. The OAuth
2.0 server is based on a third-party service, and the O2 IMS will be
registered with the OAuth server to enable OAuth functionality.

Test Plan:
1. Start O2 with the third-party OAuth server. Verify that requests
   with tokens assigned by the third-party server are authenticated
   successfully by O2 IMS.
2. Start O2 with the original K8S authentication and verify that it
   is successful.

Issue-ID: INF-447

Change-Id: I8c3cf9ce297b8404cea60c24a1d50b0fb17107a0
Signed-off-by: Kaige Sun <kaige.sun@windriver.com>
Signed-off-by: Zhang Rong(Jon) <rong.zhang@windriver.com>