Add the PATCH method for the monitoring API
In the latest specification release, a new API has been added to the
monitoring interface, which is an acknowledge/clear operation for
the alarm event.

This commit implements the O2IMS API to support SMO's requests of
alarm event acknowledgment or clearing.
The acknowledge operation will update the alarmEventRecord object,
mark the acknowledged field as True.
The clear operation will call the fault client asynchronously to
delete the alarm of the O-Cloud.

Test Plan:
1. Request the API to acknowledge/clear the alarm.
2. Request with both "alarmAcknowledged" and "perceivedSeverity"
   attribute failed since it only support one, not both.
3. Request with "perceivedSeverity" only support "5" as expected
   which is "CLEARED".
4. After SMO requests the alarm clear to IMS via IMS monitoring API,
   fault client is requested to delete the alarm from the O-Cloud.

Issue-ID: INF-449

Change-Id: I71419a9d467687d6708bfd3e9cd201dbce5e36c5
Signed-off-by: Zhang Rong(Jon) <rong.zhang@windriver.com>