Add the PATCH method for the monitoring API
Since the specification release, a new API has been added to the
monitoring interface. Which is an acknowledge/clear operation for
the alarm event.

This commit will support SMO requests the O2IMS API to
acknowledge or clear the alarm event.
The acknowledge will update the alarmEventRecord object, mark
the acknowledged as True.
The clear operation will async call the fault client to delete
the alarm of the O-Cloud.

Test Plan:
1. Request the API to acknowledge/clear the alarm.
2. Only supported one request from one of the acknowledge or
   clear.
3. The clear request should with "5" that is "CLEARED".
4. After request the alarm clear, it was requested the O-Cloud
   to delete its alarm.

Issue-ID: INF-449

Change-Id: I71419a9d467687d6708bfd3e9cd201dbce5e36c5
Signed-off-by: Zhang Rong(Jon) <rong.zhang@windriver.com>