Fix the fault client returns 500 instend of 404
The dependency on the FM module was changed, resulting in it
returning HTTP code 500 instead of HTTP code 404 for not found
errors.

This commit is a workaround to track the 500 error and skip.
When the FM module upgrade, need to be change back

Test Plan:
PASS - Started the watcher service and checked the logs on the O2
environment for alarms that do not exist on the O-Cloud.

Issue-ID: INF-457
Change-Id: Ie2050bda5dd9fc724051de5e861a82fe9f495c44
Signed-off-by: Zhang Rong(Jon) <rong.zhang@windriver.com>
(cherry picked from commit 277a26d01dba32d307cd23eef3c4f97ab1024721)