Added HTTPS support
On StarlingX openstack is now by default HTTPS.  Support must be added
for this.

Also, In the helm chart, values have been added to allowing mounting
the host cert in the watcher pod.  The mounted cert is then used to
verify the SSL certificate of the STX and DCMANAGER clients.

TEST PLAN
PASS: Run compliance and quality robot tests, with mock SMO
* With HTTPS enabled
* Use cert mounting values
* Deploy oran-o2 application on StarlingX system controller (stx9)
PASS: Run compliance and quality robot tests, with mock SMO
* Without HTTPS enabled
* Deploy oran-o2 application on StarlingX system controller (stx9)

Issue-ID: INF-461
Signed-off-by: jkraitbewr <joshua.kraitberg@windriver.com>
Change-Id: Icdd1bae303bfe29c27570c310b32d80a9f56709c