Upgrade the latest clients of O2IMS
The O2IMS request O-Cloud (StarlingX) depends on the clients from
StarlingX community.
This commit will upgrade the fault client and cgts client to the
latest to support the latest StarlingX version.
And update the local test Dockerfile, to make sure the local test
with the latest image and latest clients.

Test Plan:
1. Image build succeeded with the local test Dockerfile.
2. O2 watcher service queried the O-Cloud resource and alarm
   successfully.

Issue-ID: INF-451

Change-Id: I84236678fcef262dcd20f6bf2265b86f905d6989
Signed-off-by: Zhang Rong(Jon) <rong.zhang@windriver.com>