Fix o2pubsub container without host cert
On StarlingX openstack is now by default HTTPS. Support must be added
for this.

Also, In the helm chart, values have been added to allowing mounting
the host cert in the o2pubsub pod. The mounted cert is then used to
verify the SSL certificate of the STX and FM clients.

TEST PLAN
PASS: Run compliance and quality robot tests, with mock SMO
* With HTTPS enabled
* Use cert mounting values
* Deploy oran-o2 application on StarlingX system controller (stx10)
PASS: Run compliance and quality robot tests, with mock SMO
* Without HTTPS enabled
* Deploy oran-o2 application on StarlingX system controller (stx10)

Change-Id: I2b2e26f25e04d638a3a0814645cfa28153721b49
Signed-off-by: Zhang Rong(Jon) <rong.zhang@windriver.com>